some renaming and java docs
