+++ core/src/main/java/org/elasticsearch/indices/recovery/RecoveryTargetHandler.java
@@ -28,6 +28,12 @@ import java.util.List;
 
 
 public interface RecoveryTargetHandler {
+
+    /**
+     * Prepares the tranget to receive translog operations, after all file have been copied
+     *
+     * @param totalTranslogOps total translog operations expected to be sent
+     */
     void prepareForTranslogOperations(int totalTranslogOps) throws IOException;
 
     /**
@@ -37,23 +43,32 @@ public interface RecoveryTargetHandler {
      **/
     void finalizeRecovery();
 
+    /**
+     * Index a set of translog operations on the target
+     * @param operations operations to index
+     * @param totalTranslogOps current number of total operations expected to be indexed
+     */
     void indexTranslogOperations(List<Translog.Operation> operations, int totalTranslogOps);
 
+    /**
+     * Notifies the target of the files it is going to receive
+     */
     void receiveFileInfo(List<String> phase1FileNames,
                          List<Long> phase1FileSizes,
                          List<String> phase1ExistingFileNames,
                          List<Long> phase1ExistingFileSizes,
                          int totalTranslogOps);
 
+    /**
+     * After all source files has been sent over, this command is sent to the target so it can clean any local
+     * files that are not part of the source store
+     * @param totalTranslogOps an update number of translog operations that will be replayed later on
+     * @param sourceMetaData meta data of the source store
+     */
     void cleanFiles(int totalTranslogOps, Store.MetadataSnapshot sourceMetaData) throws IOException;
 
+    /** writes a partial file chunk to the target store */
     void writeFileChunk(StoreFileMetaData fileMetaData, long position, BytesReference content,
                         boolean lastChunk, int totalTranslogOps) throws IOException;
 
-    class RetryTranslogOpsException extends Exception {
-
-        public RetryTranslogOpsException(Throwable cause) {
-            super(cause);
-        }
-    }
 }
