+++ core/src/main/java/org/elasticsearch/indices/recovery/RemoteRecoveryTargetHandler.java
@@ -75,7 +75,7 @@ public class RemoteRecoveryTargetHandler implements RecoveryTargetHandler {
 
     @Override
     public void prepareForTranslogOperations(int totalTranslogOps) throws IOException {
-        transportService.submitRequest(targetNode, RecoveryTarget.Actions.PREPARE_TRANSLOG,
+        transportService.submitRequest(targetNode, RecoveryTargetService.Actions.PREPARE_TRANSLOG,
                 new RecoveryPrepareForTranslogOperationsRequest(recoveryId, shardId, totalTranslogOps),
                 TransportRequestOptions.builder().withTimeout(recoverySettings.internalActionTimeout()).build(),
                 EmptyTransportResponseHandler.INSTANCE_SAME).txGet();
@@ -83,7 +83,7 @@ public class RemoteRecoveryTargetHandler implements RecoveryTargetHandler {
 
     @Override
     public void finalizeRecovery() {
-        transportService.submitRequest(targetNode, RecoveryTarget.Actions.FINALIZE,
+        transportService.submitRequest(targetNode, RecoveryTargetService.Actions.FINALIZE,
                 new RecoveryFinalizeRecoveryRequest(recoveryId, shardId),
                 TransportRequestOptions.builder().withTimeout(recoverySettings.internalActionLongTimeout()).build(),
                 EmptyTransportResponseHandler.INSTANCE_SAME).txGet();
@@ -93,7 +93,7 @@ public class RemoteRecoveryTargetHandler implements RecoveryTargetHandler {
     public void indexTranslogOperations(List<Translog.Operation> operations, int totalTranslogOps) {
         final RecoveryTranslogOperationsRequest translogOperationsRequest = new RecoveryTranslogOperationsRequest(
                 recoveryId, shardId, operations, totalTranslogOps);
-        transportService.submitRequest(targetNode, RecoveryTarget.Actions.TRANSLOG_OPS, translogOperationsRequest,
+        transportService.submitRequest(targetNode, RecoveryTargetService.Actions.TRANSLOG_OPS, translogOperationsRequest,
                 translogOpsRequestOptions, EmptyTransportResponseHandler.INSTANCE_SAME).txGet();
     }
 
@@ -103,7 +103,7 @@ public class RemoteRecoveryTargetHandler implements RecoveryTargetHandler {
 
         RecoveryFilesInfoRequest recoveryInfoFilesRequest = new RecoveryFilesInfoRequest(recoveryId, shardId,
                 phase1FileNames, phase1FileSizes, phase1ExistingFileNames, phase1ExistingFileSizes, totalTranslogOps);
-        transportService.submitRequest(targetNode, RecoveryTarget.Actions.FILES_INFO, recoveryInfoFilesRequest,
+        transportService.submitRequest(targetNode, RecoveryTargetService.Actions.FILES_INFO, recoveryInfoFilesRequest,
                 TransportRequestOptions.builder().withTimeout(recoverySettings.internalActionTimeout()).build(),
                 EmptyTransportResponseHandler.INSTANCE_SAME).txGet();
 
@@ -111,7 +111,7 @@ public class RemoteRecoveryTargetHandler implements RecoveryTargetHandler {
 
     @Override
     public void cleanFiles(int totalTranslogOps, Store.MetadataSnapshot sourceMetaData) throws IOException {
-        transportService.submitRequest(targetNode, RecoveryTarget.Actions.CLEAN_FILES,
+        transportService.submitRequest(targetNode, RecoveryTargetService.Actions.CLEAN_FILES,
                 new RecoveryCleanFilesRequest(recoveryId, shardId, sourceMetaData, totalTranslogOps),
                 TransportRequestOptions.builder().withTimeout(recoverySettings.internalActionTimeout()).build(),
                 EmptyTransportResponseHandler.INSTANCE_SAME).txGet();
@@ -142,7 +142,7 @@ public class RemoteRecoveryTargetHandler implements RecoveryTargetHandler {
             throttleTimeInNanos = 0;
         }
 
-        transportService.submitRequest(targetNode, RecoveryTarget.Actions.FILE_CHUNK,
+        transportService.submitRequest(targetNode, RecoveryTargetService.Actions.FILE_CHUNK,
                 new RecoveryFileChunkRequest(recoveryId, shardId, fileMetaData, position, content, lastChunk,
                         totalTranslogOps,
                                 /* we send totalOperations with every request since we collect stats on the target and that way we can
